created basic webserver
it listens to http requests. To test a request issue "curl -X POST --data "asdasdasdasdasd" <ip>:<port>"

Also adjusted the class to accept the Config::ServerConfigStruct struct for the settings. Fixed emission
of the "ready" signal for the config class

+++ src/config.h
@@ -28,7 +28,7 @@ public:
     /** struct with server settigns */
     struct ServerConfigStruct
     {
-        int port;
+        quint16 port;
     };
 
     /** struct with agatha couchDb info */
